Fixed chunk memory usage

+++ extension/include/block.h
@@ -9,15 +9,15 @@ class Block : public Resource {
 	GDCLASS(Block, Resource)
 
 private:
-    uint64_t id = 0;
+    uint8_t id = 0;
 
 protected:
 	static void _bind_methods();
 
     void simulate(uint64_t state);
 
-    uint64_t get_id() const;
-    void set_id(uint64_t new_id);
+    uint8_t get_id() const;
+    void set_id(uint8_t new_id);
 
 public:
 	Block();
